refactor(js): move plugin definition to ai package
