+++ js/genkit/src/plugin.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
+import { type GenkitPluginV2 } from '@genkit-ai/ai';
 import { type ModelAction } from '@genkit-ai/ai/model';
 import {
   GenkitError,
-  type Action,
   type ActionMetadata,
-  type BackgroundAction,
+  type ResolvableAction,
 } from '@genkit-ai/core';
 import type { Genkit } from './genkit.js';
 import type { ActionType } from './registry.js';
@@ -32,27 +32,13 @@ export {
 } from '@genkit-ai/ai/model';
 export { reranker } from '@genkit-ai/ai/reranker';
 export { indexer, retriever } from '@genkit-ai/ai/retriever';
+export { type GenkitPluginV2, type ResolvableAction };
+
 export interface PluginProvider {
   name: string;
   initializer: () => void | Promise<void>;
   resolver?: (action: ActionType, target: string) => Promise<void>;
   listActions?: () => Promise<ActionMetadata[]>;
-}
-
-export type ResolvableAction = Action | BackgroundAction;
-
-export interface GenkitPluginV2 {
-  version: 'v2';
-  name: string;
-  init?: () => ResolvableAction[] | Promise<ResolvableAction[]>;
-  resolve?: (
-    actionType: ActionType,
-    name: string
-  ) => ResolvableAction | undefined | Promise<ResolvableAction | undefined>;
-  list?: () => ActionMetadata[] | Promise<ActionMetadata[]>;
-
-  // A shortcut for resolving a model.
-  model(name: string): Promise<ModelAction>;
 }
 
 export type GenkitPlugin = (genkit: Genkit) => PluginProvider;
